Cache digests when generating interfaces
Defines a `Digest` to compute when generating a Flow interface from a TypeScript interface, which is then written to a cache on disk. Subsequent interface generation compare the digest written in the cache to the digest of the TypeScript interface if there is already a Flow interface present. If the digests match, generating the Flow interface is skipped.

Ajv is used to define the schema of digests that are serialized to the cache and then parsed into TypeScript types.