Use digests when generating flow interfaces (#1)
* Convert to TypeScript

* Cache digests when generating interfaces

Defines a `Digest` to compute when generating a Flow interface from a TypeScript interface, which is then written to a cache on disk. Subsequent interface generation compare the digest written in the cache to the digest of the TypeScript interface if there is already a Flow interface present. If the digests match, generating the Flow interface is skipped.

Ajv is used to define the schema of digests that are serialized to the cache and then parsed into TypeScript types.

* Build

* Bump version

* Restyled by prettier (#2)

Co-authored-by: Restyled.io <commits@restyled.io>

* Add formatting and tests

* Build

* Restyled by prettier (#3)

Co-authored-by: Restyled.io <commits@restyled.io>

---------

Co-authored-by: restyled-io[bot] <32688539+restyled-io[bot]@users.noreply.github.com>
Co-authored-by: Restyled.io <commits@restyled.io>